result count in detail report

+++ src/main/java/com/example/starter/service/ResultService.java
@@ -34,7 +34,7 @@ public class ResultService extends BaseService {
         }
     }
 
-    public static List<Result> findAllByTestRunId(int testRunId) {
+    public static List<Result> findAllByTestRunId(long testRunId) {
         List<Result> result = new ArrayList<>();
         EntityManager em = getEntityManager();
         try {
